+++ src/lib/auth/Authentication.ts
@@ -1,8 +1,7 @@
-import type { ScholarID } from '$lib/types/Scholar';
-
-/** An interface that defaults authentication functionality. */
-export default abstract class Authentication {
-	abstract getScholarID(): ScholarID;
-
-	abstract logout(): void;
+/** An interface that defines an authentication state interface. Useful if we want to migrate to other auth providers. */
+export default abstract class Authentication<UserKind> {
+	abstract setUser(user: UserKind | null): void;
+	abstract getUserID(): string | null;
+	abstract isAuthenticated(): boolean;
+	abstract signOut(): void;
 }