+++ src/routes/+layout.svelte
@@ -1,28 +1,21 @@
-<script lang="ts" context="module">
-	const AuthContext = Symbol('auth');
-
-	export function getAuth() {
-		return getContext<Auth>(AuthContext);
-	}
-</script>
-
 <script lang="ts">
 	import { dev } from '$app/environment';
 	import { goto } from '$app/navigation';
 	import Header from '$lib/components/Header.svelte';
 	import { AuthSymbol, DatabaseSymbol } from '$lib/Context';
 	import MockDatabase from '$lib/data/MockDatabase';
-	import { getContext, onMount, setContext, type Snippet } from 'svelte';
+	import { onMount, setContext, type Snippet } from 'svelte';
 	import { invalidate } from '$app/navigation';
-	import type { Session, SupabaseClient, User } from '@supabase/supabase-js';
-	import Auth from './Auth.svelte';
+	import type { Session, SupabaseClient } from '@supabase/supabase-js';
+	import { createAuthContext, getAuth } from './Auth.svelte';
 
 	let {
 		data,
 		children
 	}: { data: { session: Session; supabase: SupabaseClient }; children: Snippet } = $props();
 
-	let auth = new Auth(data.supabase);
+	createAuthContext(data.supabase);
+	const auth = getAuth();
 
 	/** Always go home in production, pre-release */
 	onMount(() => {
@@ -41,11 +34,6 @@
 
 	/** Set a database connection context for all to use. */
 	setContext(DatabaseSymbol, new MockDatabase());
-
-	/** Set an auth connection context for all to use */
-	setContext(AuthSymbol, auth);
-
-	setContext(AuthContext, auth);
 </script>
 
 {#if dev}