+++ src/routes/Auth.svelte.ts
@@ -1,10 +1,14 @@
+import Authentication from '$lib/auth/Authentication';
 import type { SupabaseClient, User } from '@supabase/supabase-js';
+import { getContext, setContext } from 'svelte';
 
-export default class Auth {
-	user = $state<User | null>(null);
-	client: SupabaseClient;
+/** Represents the current authenatication state from Supabase. */
+export default class SupabaseAuth extends Authentication<User> {
+	private user = $state<User | null>(null);
+	private client: SupabaseClient;
 
 	constructor(supabase: SupabaseClient) {
+		super();
 		this.client = supabase;
 	}
 
@@ -12,12 +16,16 @@ export default class Auth {
 		this.user = user;
 	}
 
-	authenticated() {
+	isAuthenticated() {
 		return this.user !== null;
 	}
 
-	signOut() {
-		return this.client.auth.signOut();
+	getUserID(): string | null {
+		return this.user?.id ?? null;
+	}
+
+	async signOut() {
+		return await this.client.auth.signOut();
 	}
 
 	async signIn(email: string) {
@@ -26,4 +34,14 @@ export default class Auth {
 		});
 		return error?.message;
 	}
+}
+
+const AuthSymbol = Symbol('auth');
+
+export function createAuthContext(supabase: SupabaseClient) {
+	setContext(AuthSymbol, new SupabaseAuth(supabase));
+}
+
+export function getAuth(): Authentication<User> {
+	return getContext<SupabaseAuth>(AuthSymbol);
 }