+++ src/routes/login/+page.svelte
@@ -2,17 +2,14 @@
 	import Button from '$lib/components/Button.svelte';
 	import TextField from '$lib/components/TextField.svelte';
 	import { ORCIDRegex } from '$lib/types/Scholar';
-	import { getAuth } from '../+layout.svelte';
 	import Status from '$lib/components/Status.svelte';
+	import Link from '$lib/components/Link.svelte';
+	import { getAuth } from '../Auth.svelte';
 
-	let email = $state('');
+	let orcid = $state('');
 
 	function validORCID(text: string) {
-		return /** TODO For testing */ text.length === 0 || ORCIDRegex.test(text);
-	}
-
-	function validEmail(email: string) {
-		return /^\S+@\S+\.\S+$/.test(email);
+		return text.length === 0 || ORCIDRegex.test(text);
 	}
 
 	let auth = getAuth();
@@ -20,30 +17,26 @@
 	let error = $state<undefined | string>(undefined);
 	let submitted = $state(false);
 
-	async function login() {
-		error = await auth.signIn(email);
-		if (!error) submitted = true;
-	}
+	async function login() {}
 </script>
 
 <h1>Login</h1>
 
-{#if auth.authenticated()}
+{#if auth}
 	<p>You are logged in.</p>
 {:else}
-	<p>ORCID logins are coming. For now, enter your email to receive a one-time password.</p>
-	<!-- <p>Login with your <Link to="https://orcid.org/">ORCID</Link> account.</p> -->
+	<p>Login with your <Link to="https://orcid.org/">ORCID</Link> account.</p>
 
 	<p>
 		<TextField
 			active={!submitted}
 			name="email"
 			size={19}
-			bind:text={email}
+			bind:text={orcid}
 			placeholder="email"
-			valid={validEmail}
+			valid={validORCID}
 		/>
-		<Button action={() => login()} active={!submitted && validEmail(email)}>Login</Button>
+		<Button action={() => login()} active={!submitted && validORCID(orcid)}>Login</Button>
 	</p>
 {/if}